requirements.txt
mwclient
discord.py